Reorganize prop forwarding code into multiple files, improve naming
