--- tools/analyzer_plugin/lib/src/util/forwarded_props.dart
@@ -1,299 +1,0 @@
-import 'package:analyzer/dart/ast/ast.dart';
-import 'package:analyzer/dart/element/element.dart';
-import 'package:analyzer/dart/element/type.dart';
-import 'package:over_react_analyzer_plugin/src/diagnostic_contributor.dart';
-import 'package:over_react_analyzer_plugin/src/util/ast_util.dart';
-import 'package:over_react_analyzer_plugin/src/util/is_props_from_render.dart';
-import 'package:over_react_analyzer_plugin/src/util/util.dart';
-
-ForwardedProps? getForwardedProps(FluentComponentUsage usage) {
-  late final enclosingComponentPropsClass =
-      getTypeOfPropsInEnclosingInterface(usage.node)?.typeOrBound.element.tryCast<InterfaceElement>();
-  late final enclosingComponentForwardedProps = _getForwardedPropsFromEnclosingInterface(usage.node);
-
-  for (final invocation in usage.cascadedMethodInvocations) {
-    final methodName = invocation.methodName.name;
-    final arg = invocation.node.argumentList.arguments.firstOrNull;
-
-    // FIXME what should we do about the second condition arg for addProps?
-    final isAddAllOrAddProps = methodName == 'addProps' || methodName == 'addAll';
-
-    // ..addProps(props)
-    if (isAddAllOrAddProps && arg != null && isPropsFromRender(arg)) {
-      final propsType = arg.staticType?.typeOrBound.tryCast<InterfaceType>()?.element;
-      if (propsType != null) {
-        return ForwardedProps(propsType, PropsToForward.all(), invocation.node);
-      }
-    } else if (
-        // ..addProps(props.getPropsToForward(...))
-        (isAddAllOrAddProps && arg is MethodInvocation && arg.methodName.name == 'getPropsToForward') ||
-            // ..modifyProps(props.addPropsToForward(...))
-            (methodName == 'modifyProps' && arg is MethodInvocation && arg.methodName.name == 'addPropsToForward')) {
-      final realTarget = arg.realTarget;
-      if (realTarget != null && isPropsFromRender(realTarget)) {
-        final propsType = realTarget.staticType?.typeOrBound.tryCast<InterfaceType>()?.element;
-        if (propsType != null) {
-          return ForwardedProps(propsType, _parseGetPropsToForward(arg.argumentList, propsType), invocation.node);
-        }
-      }
-    } else if (
-        // ..addProps(copyUnconsumedProps())
-        (isAddAllOrAddProps && arg is MethodInvocation && arg.methodName.name == 'copyUnconsumedProps') ||
-            // ..modifyProps(addUnconsumedProps)
-            (methodName == 'modifyProps' && arg is Identifier && arg.name == 'addUnconsumedProps')) {
-      if (enclosingComponentPropsClass != null && enclosingComponentForwardedProps != null) {
-        return ForwardedProps(enclosingComponentPropsClass, enclosingComponentForwardedProps, invocation.node);
-      }
-    } else if (
-        // ..addUnconsumedProps(props, consumedProps)
-        methodName == 'addUnconsumedProps') {
-      final consumedPropsArg = invocation.node.argumentList.arguments.elementAtOrNull(1);
-      if (arg != null && consumedPropsArg != null && isPropsFromRender(arg)) {
-        final propsType = arg.staticType?.typeOrBound.tryCast<InterfaceType>()?.element;
-        if (propsType != null) {
-          final propsToForward = _getForwardedPropsFromConsumedProps(consumedPropsArg);
-          return ForwardedProps(propsType, propsToForward, invocation.node);
-        }
-      }
-    }
-  }
-
-  return null;
-}
-
-class ForwardedProps {
-  final InterfaceElement propsClassBeingForwarded;
-  final PropsToForward propsToForward;
-  final AstNode debugSourceNode;
-
-  ForwardedProps(this.propsClassBeingForwarded, this.propsToForward, this.debugSourceNode);
-
-  bool definitelyForwardsPropsFrom(InterfaceElement propsClass) =>
-      !propsToForward.mightExcludeClass(propsClass) &&
-      propsClassBeingForwarded.allSupertypes.any((s) => s.element == propsClass);
-
-  // TODO loop through props mixins and show the full list of props being forwarded for debug purposes
-  @override
-  String toString() => 'Props from ${propsClassBeingForwarded.name}: $propsToForward';
-}
-
-abstract class PropsToForward {
-  const PropsToForward();
-
-  const factory PropsToForward.all() = _PropsToForward$All;
-
-  const factory PropsToForward.allExceptFor(Set<InterfaceElement> onlyProps) = _PropsToForward$AllExceptFor;
-
-  const factory PropsToForward.only(Set<InterfaceElement> excludedProps) = _PropsToForward$Only;
-
-  const factory PropsToForward.unresolved() = _PropsToForward$Unresolved;
-
-  bool mightExcludeClass(InterfaceElement e);
-
-  String get debugDescription;
-
-  @override
-  toString() => '$debugDescription';
-}
-
-class _PropsToForward$All extends PropsToForward {
-  const _PropsToForward$All();
-
-  @override
-  bool mightExcludeClass(InterfaceElement e) => false;
-
-  @override
-  String get debugDescription => 'all';
-}
-
-class _PropsToForward$Only extends PropsToForward {
-  final Set<InterfaceElement> _onlyProps;
-
-  const _PropsToForward$Only(this._onlyProps);
-
-  @override
-  bool mightExcludeClass(InterfaceElement e) => !_onlyProps.contains(e);
-
-  @override
-  String get debugDescription => 'only ${_onlyProps.map((e) => e.name).toSet()}';
-}
-
-class _PropsToForward$AllExceptFor extends PropsToForward {
-  final Set<InterfaceElement> _excludedProps;
-
-  const _PropsToForward$AllExceptFor(this._excludedProps);
-
-  @override
-  bool mightExcludeClass(InterfaceElement e) => _excludedProps.contains(e);
-
-  @override
-  String get debugDescription => 'all except ${_excludedProps.map((e) => e.name).toSet()}';
-}
-
-class _PropsToForward$Unresolved extends PropsToForward {
-  const _PropsToForward$Unresolved();
-
-  @override
-  bool mightExcludeClass(InterfaceElement e) => false;
-
-  @override
-  String get debugDescription => 'unresolved';
-}
-
-PropsToForward _parseGetPropsToForward(ArgumentList argumentList, InterfaceElement propsType) {
-  final excludeArg = argumentList.arguments
-      .whereType<NamedExpression>()
-      .firstWhereOrNull((a) => a.name.label.name == 'exclude')
-      ?.expression;
-
-  // With no exclude, the default is the props in the current props type
-  if (excludeArg == null || excludeArg is NullLiteral) {
-    return PropsToForward.allExceptFor({propsType});
-  }
-
-  // Not a literal we can statically analyze
-  if (excludeArg is! SetOrMapLiteral || !excludeArg.isSet) {
-    return const PropsToForward.unresolved();
-  }
-
-  final excluded = lookUpInterfaces(excludeArg.elements)?.toSet();
-  if (excluded == null) return const PropsToForward.unresolved();
-  return PropsToForward.allExceptFor(excluded);
-}
-
-List<InterfaceElement>? lookUpInterfaces(List<CollectionElement> elements) {
-  final interfaces = <InterfaceElement>[];
-  for (final setElement in elements) {
-    // Some other element or expression we can't statically analyze.
-    // It seems like this should be a TypeLiteral,
-    // but it seems to be a SimpleIdentifier.
-    // Handle both cases in case it's sometimes one and sometimes the other.
-    InterfaceElement? interface;
-    if (setElement is Identifier) {
-      interface = setElement.staticElement.tryCast();
-    } else if (setElement is TypeLiteral) {
-      interface = setElement.type.element.tryCast();
-    }
-    if (interface == null) return null;
-
-    interfaces.add(interface);
-  }
-  return interfaces;
-}
-
-PropsToForward _getForwardedPropsFromConsumedProps(Expression consumedProps) {
-  // Look up value of consumedProps stored into variables, which is very common in usage:
-  //
-  //    final consumedProps = props.staticMeta.forMixins({...});
-  if (consumedProps is Identifier) {
-    final staticElement = consumedProps.staticElement;
-    if (staticElement == null) return PropsToForward.unresolved();
-
-    final variableValue = lookUpVariable(staticElement, consumedProps.root)?.initializer;
-    // Don't recurse for Identifiers, to prevent potential infinite loops;
-    // this case should not be common so we're safe to ignore it.
-    if (variableValue != null && variableValue is! Identifier) {
-      return _getForwardedPropsFromConsumedProps(variableValue);
-    }
-
-    return PropsToForward.unresolved();
-  }
-
-  // We could validate the target of this call, but these methods are specifically-named enough that false positives
-  // are unlikely, and validating the target would require a bit of extra logic.
-  //
-  // This invocation will usually look like:
-  // - `propsMeta.forMixins` - class component `propsMeta` getter
-  // - `props.staticMeta.forMixins` - props class `staticMeta` getter:
-  if (consumedProps is MethodInvocation) {
-    if (consumedProps.methodName.name == 'forMixins') {
-      final arg = consumedProps.argumentList.arguments.whereType<Expression>().firstOrNull;
-      if (arg is SetOrMapLiteral) {
-        final mixins = lookUpInterfaces(arg.elements)?.toSet();
-        if (mixins != null) {
-          return PropsToForward.allExceptFor(mixins);
-        }
-      }
-      return PropsToForward.unresolved();
-    } else if (consumedProps.methodName.name == 'allExceptForMixins') {
-      final arg = consumedProps.argumentList.arguments.whereType<Expression>().firstOrNull;
-      if (arg is SetOrMapLiteral) {
-        final mixins = lookUpInterfaces(arg.elements)?.toSet();
-        if (mixins != null) {
-          // FIXME this also includes props not in this props class; how to represent that clearnly
-          return PropsToForward.only(mixins);
-        }
-      }
-      return PropsToForward.unresolved();
-    }
-  }
-
-  final consumedPropElements =
-      consumedProps.tryCast<ListLiteral>()?.elements ?? consumedProps.tryCast<SetOrMapLiteral>()?.elements;
-  if (consumedPropElements != null) {
-    // get consumedProps => const [LightboxProps.meta]
-    final excluded = <InterfaceElement>{};
-    for (final consumedPropElement in consumedPropElements) {
-      if (consumedPropElement is Expression) {
-        final result = getSimpleTargetAndPropertyName(consumedPropElement, allowMethodInvocation: false);
-        if (result != null && result.item2.name != 'meta') {
-          final propsClass = result.item1.staticElement.tryCast<InterfaceElement>();
-          if (propsClass != null) {
-            excluded.add(propsClass);
-            continue;
-          }
-        }
-      }
-      // Short-circuit if we encounter any other cases.
-      return const PropsToForward.unresolved();
-    }
-    return PropsToForward.allExceptFor(excluded);
-  }
-
-  return PropsToForward.unresolved();
-}
-
-PropsToForward? _getForwardedPropsFromEnclosingInterface(AstNode node) {
-  final enclosingComponentPropsClass =
-      getTypeOfPropsInEnclosingInterface(node)?.typeOrBound.element.tryCast<InterfaceElement>();
-  if (enclosingComponentPropsClass == null) return null;
-
-  final enclosingClass = node.thisOrAncestorOfType<ClassDeclaration>();
-  final consumedPropsGetter =
-      enclosingClass?.members.whereType<MethodDeclaration>().firstWhereOrNull((m) => m.name.lexeme == 'consumedProps');
-
-  // FIXME double check this behavior.
-  // The default is only props declared in enclosingComponentPropsClass for legacy boilerplate, and empty for non-legacy boilerplate.
-  if (consumedPropsGetter == null) {
-    // FIXME implement conditionally for legacy boilerplate
-    // ignore: prefer_const_declarations
-    final isLegacy = false;
-    // ignore: dead_code
-    return isLegacy ? PropsToForward.allExceptFor({enclosingComponentPropsClass}) : PropsToForward.only({});
-  }
-
-  final consumedProps = consumedPropsGetter.body.returnExpressions.singleOrNull;
-  if (consumedProps == null) {
-    return PropsToForward.unresolved();
-  }
-
-  return _getForwardedPropsFromConsumedProps(consumedProps);
-}
-
-DartType? getTypeOfPropsInEnclosingInterface(AstNode node) {
-  final enclosingInterface = getEnclosingInterface(node);
-  if (enclosingInterface == null) return null;
-
-  return enclosingInterface.thisType.lookUpGetter2('props', enclosingInterface.library)?.returnType;
-}
-
-InterfaceElement? getEnclosingInterface(AstNode node) {
-  for (final ancestor in node.ancestors) {
-    if (ancestor is ClassDeclaration) return ancestor.declaredElement;
-    if (ancestor is MixinDeclaration) return ancestor.declaredElement;
-    if (ancestor is ExtensionDeclaration) {
-      return ancestor.declaredElement?.extendedType.typeOrBound.tryCast<InterfaceType>()?.element;
-    }
-  }
-  return null;
-}
